+++ android/src/main/java/com/barkoder_flutter/Util.java
@@ -8,68 +8,95 @@ import android.util.Base64;
 import com.barkoder.Barkoder;
 import com.barkoder.BarkoderLog;
 
+import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
 
 import java.io.ByteArrayOutputStream;
+import java.util.Objects;
 
 class Util {
     private static final String TAG = Util.class.getSimpleName();
 
     static String barkoderResultsToJsonString(Barkoder.Result[] results, Bitmap[] thumbnails, Bitmap resultImage) {
-        JSONObject resultJson = new JSONObject();
+        JSONObject barkoderResultJson = new JSONObject();
+        JSONArray resultsArray = new JSONArray();
 
         try {
-            resultJson.put("barcodeType", results[0].barcodeType.ordinal());
-            resultJson.put("barcodeTypeName", results[0].barcodeTypeName);
-            resultJson.put("binaryDataAsBase64", Base64.encodeToString(results[0].binaryData, Base64.NO_WRAP));
-            resultJson.put("textualData", results[0].textualData);
+            // Process each decoder result separately
+            for (Barkoder.Result decoderResult : results) {
+                JSONObject resultJson = new JSONObject();
 
-            if (!TextUtils.isEmpty(results[0].characterSet))
-                resultJson.put("characterSet", results[0].characterSet);
+                resultJson.put("barcodeType", decoderResult.barcodeType.ordinal());
+                resultJson.put("barcodeTypeName", decoderResult.barcodeTypeName);
+                resultJson.put("binaryDataAsBase64", Base64.encodeToString(decoderResult.binaryData, Base64.NO_WRAP));
+                resultJson.put("textualData", decoderResult.textualData);
 
-            if (results[0].extra != null && results[0].extra.length > 0) {
-                JSONObject extraAsJson = new JSONObject();
-                for (Barkoder.BKKeyValue item : results[0].extra) {
-                    extraAsJson.put(item.key, item.value);
+                if (!TextUtils.isEmpty(decoderResult.characterSet)) {
+                    resultJson.put("characterSet", decoderResult.characterSet);
                 }
-                resultJson.put("extra", extraAsJson.toString());
-            }
 
-            if (resultImage != null)
-                resultJson.put("resultImageAsBase64", bitmapImageToBase64(resultImage));
+                // Add "extra" if available and not empty
+                if (decoderResult.extra != null && decoderResult.extra.length > 0) {
+                    JSONObject extraJson = new JSONObject();
+                    for (Barkoder.BKKeyValue item : decoderResult.extra) {
+                        extraJson.put(item.key, item.value);
+                    }
+                    resultJson.put("extra", extraJson.toString());
+                }
 
-            if (thumbnails[0] != null)
-                resultJson.put("resultThumbnailAsBase64", bitmapImageToBase64(thumbnails[0]));
+                // Add mrzImagesAsBase64
+                if (Objects.equals(decoderResult.barcodeTypeName, "MRZ")) {
+                    if (decoderResult.images != null) {
+                        JSONArray mrzImagesArray = new JSONArray();
 
-            if (results[0].images != null) {
-                for (Barkoder.BKImageDescriptor image : results[0].images) {
-                    if (image != null && image.image != null) {
-                        switch (image.name) {
-                            case "main":
-                                resultJson.put("mainImageAsBase64", bitmapImageToBase64(image.image));
-                                break;
-                            case "document":
-                                resultJson.put("documentImageAsBase64", bitmapImageToBase64(image.image));
-                                break;
-                            case "signature":
-                                resultJson.put("signatureImageAsBase64", bitmapImageToBase64(image.image));
-                                break;
-                            case "picture":
-                                resultJson.put("pictureImageAsBase64", bitmapImageToBase64(image.image));
-                                break;
-                            default:
-                                break;
+                        for (Barkoder.BKImageDescriptor image : decoderResult.images) {
+                            if (image != null && image.image != null) {
+                                switch (image.name) {
+                                    case "main":
+                                    case "document":
+                                    case "signature":
+                                    case "picture":
+                                        JSONObject imageInfo = new JSONObject();
+                                        imageInfo.put("name", image.name);
+                                        imageInfo.put("base64", bitmapImageToBase64(image.image));
+                                        mrzImagesArray.put(imageInfo);
+                                        break;
+                                }
+                            }
                         }
+                        resultJson.put("mrzImagesAsBase64", mrzImagesArray);
                     }
                 }
+
+                resultsArray.put(resultJson);
             }
+
+            // Add decoderResults to the final JSON object
+            barkoderResultJson.put("decoderResults", resultsArray);
+
+            // Process thumbnails as an array of base64 strings if available, outside the loop
+            if (thumbnails != null) {
+                JSONArray thumbnailsBase64Array = new JSONArray();
+                for (Bitmap thumbnail : thumbnails) {
+                    if (thumbnail != null) {
+                        thumbnailsBase64Array.put(bitmapImageToBase64(thumbnail));
+                    }
+                }
+                barkoderResultJson.put("resultThumbnailsAsBase64", thumbnailsBase64Array);
+            }
+
+            // Process the main result image as base64 if available, outside the loop
+            if (resultImage != null) {
+                barkoderResultJson.put("resultImageAsBase64", bitmapImageToBase64(resultImage));
+            }
+
 
         } catch (JSONException ex) {
             BarkoderLog.d(TAG, ex.getMessage());
         }
 
-        return resultJson.toString();
+        return barkoderResultJson.toString();
     }
 
     private static String bitmapImageToBase64(Bitmap bitmapImage) {