+++ lib/barkoder_flutter_models.dart
@@ -1,95 +1,130 @@
 import 'dart:convert';
+import 'dart:typed_data';
 
 class BarkoderResult {
-  late BarcodeType barcodeType;
+  List<DecoderResult> decoderResults;
+  List<Uint8List>? resultThumbnails;
+  Uint8List? resultImage;
+
+  BarkoderResult({
+    required this.decoderResults,
+    this.resultThumbnails,
+    this.resultImage,
+  });
+
+  factory BarkoderResult.fromJson(Map<String, dynamic> json) {
+    List<DecoderResult> decoderResults = (json['decoderResults'] as List)
+        .map((result) => DecoderResult.fromJson(result))
+        .toList();
+    return BarkoderResult(
+      decoderResults: decoderResults,
+      resultThumbnails: json['resultThumbnailsAsBase64'] != null
+          ? (json['resultThumbnailsAsBase64'] as List<dynamic>)
+              .map((thumbnail) => Base64Decoder().convert(thumbnail))
+              .toList()
+          : null,
+      resultImage: json['resultImageAsBase64'] != null
+          ? Base64Decoder().convert(json['resultImageAsBase64'])
+          : null,
+    );
+  }
+
+  @override
+  String toString() {
+    return '{$decoderResults, $resultThumbnails, $resultImage}';
+  }
+}
+
+class DecoderResult {
+  late int barcodeType;
   late String barcodeTypeName;
   late String binaryDataAsBase64;
   late String textualData;
   String? characterSet;
   Map<String, dynamic>? extra;
-  String? resultImageAsBase64;
-  String? resultThumbnailAsBase64;
-  String? mainImageAsBase64;
-  String? documentImageAsBase64;
-  String? signatureImageAsBase64;
-  String? pictureImageAsBase64;
+  List<MRZImage>? mrzImages;
 
-  BarkoderResult(
-      {required this.barcodeType,
-      required this.barcodeTypeName,
-      required this.binaryDataAsBase64,
-      required this.textualData,
-      this.characterSet,
-      this.extra,
-      this.resultImageAsBase64,
-      this.resultThumbnailAsBase64,
-      this.mainImageAsBase64,
-      this.documentImageAsBase64,
-      this.signatureImageAsBase64,
-      this.pictureImageAsBase64});
+  DecoderResult({
+    required this.barcodeType,
+    required this.barcodeTypeName,
+    required this.binaryDataAsBase64,
+    required this.textualData,
+    this.characterSet,
+    this.extra,
+    this.mrzImages,
+  });
 
-  BarkoderResult.fromJsonString(String jsonString) {
-    Map<String, dynamic> resultMap = json.decode(jsonString);
-
-    int barcodeTypeIndex = resultMap['barcodeType'];
-    barcodeType = BarcodeType.fromInt(barcodeTypeIndex);
+  DecoderResult.fromJson(Map<String, dynamic> resultMap) {
+    barcodeType = resultMap['barcodeType'];
     barcodeTypeName = resultMap['barcodeTypeName'];
     binaryDataAsBase64 = resultMap['binaryDataAsBase64'];
     textualData = resultMap['textualData'];
     characterSet = resultMap['characterSet'];
     if (resultMap.containsKey('extra')) extra = json.decode(resultMap['extra']);
-    resultImageAsBase64 = resultMap['resultImageAsBase64'];
-    resultThumbnailAsBase64 = resultMap['resultThumbnailAsBase64'];
-    mainImageAsBase64 = resultMap['mainImageAsBase64'];
-    documentImageAsBase64 = resultMap['documentImageAsBase64'];
-    signatureImageAsBase64 = resultMap['signatureImageAsBase64'];
-    pictureImageAsBase64 = resultMap['pictureImageAsBase64'];
+    if (resultMap.containsKey('mrzImagesAsBase64')) {
+      mrzImages = (resultMap['mrzImagesAsBase64'] as List<dynamic>)
+          .map((imageData) => MRZImage.fromJson(imageData))
+          .toList();
+    }
+  }
+
+  static List<DecoderResult> fromJsonString(String jsonString) {
+    List<dynamic> resultList = json.decode(jsonString);
+    return resultList
+        .map((resultMap) => DecoderResult.fromJson(resultMap))
+        .toList();
   }
 
   @override
   String toString() {
-    return '{$barcodeType, $barcodeTypeName, $binaryDataAsBase64, $textualData, $characterSet, $extra, $resultImageAsBase64, $resultThumbnailAsBase64}';
+    return '{$barcodeType, $barcodeTypeName, $binaryDataAsBase64, $textualData, $characterSet, $extra, $mrzImages';
+  }
+}
+
+class MRZImage {
+  final String name;
+  final Uint8List value;
+
+  MRZImage({required this.name, required this.value});
+
+  factory MRZImage.fromJson(Map<String, dynamic> json) {
+    return MRZImage(
+      name: json['name'] as String,
+      value: base64Decode(json['base64'] as String),
+    );
   }
 }
 
 enum BarcodeType {
-  aztec(0),
-  aztecCompact(1),
-  qr(2),
-  qrMicro(3),
-  code128(4),
-  code93(5),
-  code39(6),
-  codabar(7),
-  code11(8),
-  msi(9),
-  upcA(10),
-  upcE(11),
-  upcE1(12),
-  ean13(13),
-  ean8(14),
-  pdf417(15),
-  pdf417Micro(16),
-  datamatrix(17),
-  code25(18),
-  interleaved25(19),
-  itf14(20),
-  iata25(21),
-  matrix25(22),
-  datalogic25(23),
-  coop25(24),
-  code32(25),
-  telepen(26),
-  dotcode(27),
-  idDocument(29);
-
-  final int value;
-  const BarcodeType(this.value);
-
-  static BarcodeType fromInt(int value) {
-    return BarcodeType.values.firstWhere((e) => e.value == value,
-        orElse: () => throw RangeError('Invalid barcodeType value: $value'));
-  }
+  aztec,
+  aztecCompact,
+  qr,
+  qrMicro,
+  code128,
+  code93,
+  code39,
+  codabar,
+  code11,
+  msi,
+  upcA,
+  upcE,
+  upcE1,
+  ean13,
+  ean8,
+  pdf417,
+  pdf417Micro,
+  datamatrix,
+  code25,
+  interleaved25,
+  itf14,
+  iata25,
+  matrix25,
+  datalogic25,
+  coop25,
+  code32,
+  telepen,
+  dotcode,
+  idDocument;
 }
 
 enum FormattingType { disabled, automatic, gs1, aamva }
@@ -108,7 +143,7 @@ enum Code39ChecksumType { disabled, enabled }
 
 enum Code11ChecksumType { disabled, single, double }
 
-enum DecodingSpeed { fast, normal, slow }
+enum DecodingSpeed { fast, normal, slow, rigorous }
 
 enum BarkoderResolution { normal, high }
 
@@ -194,8 +229,8 @@ class BarkoderConfig {
 class DekoderConfig {
   BarcodeConfig? aztec;
   BarcodeConfig? aztecCompact;
-  BarcodeConfig? qr;
-  BarcodeConfig? qrMicro;
+  BarcodeConfigWithDpmMode? qr;
+  BarcodeConfigWithDpmMode? qrMicro;
   BarcodeConfigWithLength? code128;
   BarcodeConfigWithLength? code93;
   Code39BarcodeConfig? code39;
@@ -209,7 +244,7 @@ class DekoderConfig {
   BarcodeConfig? ean8;
   BarcodeConfig? pdf417;
   BarcodeConfig? pdf417Micro;
-  DatamatrixBarcodeConfig? datamatrix;
+  BarcodeConfigWithDpmMode? datamatrix;
   BarcodeConfig? code25;
   BarcodeConfig? interleaved25;
   BarcodeConfig? itf14;
@@ -220,7 +255,7 @@ class DekoderConfig {
   BarcodeConfig? code32;
   BarcodeConfig? telepen;
   BarcodeConfig? dotcode;
-  BarcodeConfig? idDocument;
+  IdDocumentBarcodeConfig? idDocument;
   GeneralSettings? general;
 
   DekoderConfig(
@@ -435,15 +470,15 @@ class Code11BarcodeConfig {
   }
 }
 
-class DatamatrixBarcodeConfig {
+class BarcodeConfigWithDpmMode {
   bool? enabled;
   int? dpmMode;
   int? _minLength;
   int? _maxLength;
 
-  DatamatrixBarcodeConfig({this.enabled, this.dpmMode});
+  BarcodeConfigWithDpmMode({this.enabled, this.dpmMode});
 
-  DatamatrixBarcodeConfig.setLengthRange(
+  BarcodeConfigWithDpmMode.setLengthRange(
       {this.enabled, required minLength, required maxLength})
       : _minLength = minLength,
         _maxLength = maxLength;
@@ -464,6 +499,26 @@ class DatamatrixBarcodeConfig {
   setLengthRange(int minLength, int maxLength) {
     _minLength = minLength;
     _maxLength = maxLength;
+  }
+}
+
+enum IdDocumentMasterChecksumType { disabled, enabled }
+
+class IdDocumentBarcodeConfig {
+  bool? enabled;
+  IdDocumentMasterChecksumType? masterChecksum;
+
+  IdDocumentBarcodeConfig({this.enabled, this.masterChecksum});
+
+  Map<String, dynamic> toMap() {
+    Map<String, dynamic> filteredMap = {
+      "enabled": enabled,
+      "masterChecksum": masterChecksum?.index
+    };
+
+    filteredMap.removeWhere((key, value) => value == null);
+
+    return filteredMap;
   }
 }